fix: MDF 패널도 PET 코팅 시 +0.5mm 적용 및 두께 파싱 오류 수정
- 백패널/서랍바닥(MDF): PET 코팅 시 9→9.5mm, 5→5.5mm
- CNC 옵티마이저 그룹화 키 파싱: parseInt→parseFloat (소수점 두께 손실 방지)
- parseInt("9.5")=9, parseInt("15.5")=15 → parseFloat로 정확한 값 유지

+++ src/editor/CNCOptimizer/CNCOptimizerPro.tsx
@@ -771,13 +771,13 @@ function PageInner(){
         let thickness: number;
         
         if (key.startsWith('THICKNESS_')) {
-          // 재질 구분 없이 두께만 고려: THICKNESS_18
-          thickness = parseInt(key.split('_')[1]) || 18;
+          // 재질 구분 없이 두께만 고려: THICKNESS_18.5
+          thickness = parseFloat(key.split('_')[1]) || 18;
         } else {
-          // 재질+두께: PET_18
+          // 재질+두께: PET_18.5
           const parts = key.split('_');
           material = parts[0];
-          thickness = parseInt(parts[1]) || 18;
+          thickness = parseFloat(parts[1]) || 18;
         }
         
         // Find matching stock by material and thickness
